requirements.txt
flask
numpy
pillow
joblib
scikit-learn
tensorflow
tqdm
